refactor: page 폴더 기본 파일들 생성, MobileLayout 삭제
