+++ src/components/pages/stats/EmotionStats.tsx
